Various updates
I got sloppy and started making lots of different changes without
focusing or committing along the way. The main functionality changes
are (1) the student is now presented with forms to log yesterday _and_
today if either hasn't been logged, and (2) students can only update
notes for dates in the past of the current month.

Behind the scenes, I rewrote quite a bit more. I changed where the
post code was getting processed to get it out out `/`. I added a bunch
more data fields to the calendar HTML and tried to use those for
running the forms and some of the styles. There were a lot of other
tweaks too, but I've lost track of them all.